Make CUDA GDR auto-detect the mapping from GPUs to InfiniBand NICs (#258)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/258

GPU machines with InfiniBand often have several NICs, and GPUs need to be carefully mapped to them to avoid bottlenecks and congestion on the PCI links. Until now, for simplicity, we asked the user to provide such mapping when constructing the context, but this of course isn't practical. In this diff I introduce auto-detection for that mapping, by associating a GPU with the NIC that shares the longest prefix in the PCI "path".

I'm keeping the option to override this auto-detection and still manually specify a mapping, as a sort of "killswitch" should this logic not work for everyone.

Reviewed By: beauby

Differential Revision: D25824239

fbshipit-source-id: 190254c342783e6f5facca7ca1717b863b08168d