Mark cuda_ipc viable only if all devices have peer access available.
Since `cudaIpcOpenMemHandle()` will lazily try to enable peer access,
we must ensure it will not fail.
One limitation of this check is that it can be fooled by the
`CUDA_VISIBLE_DEVICES` environment variable, as two processes might be
each limited to a viable set of devices, while pairs across those sets
might not be able to access each other. One option would be to use an
ordered list of device UUIDS as part of the domain descriptor (those
would need to be accessed through the `nvmlDeviceGetUUID()` function
from the nvml library). A simpler option would be to check the
process' environment and mark the channel non-viable if
`CUDA_VISIBLE_DEVICES` is set.