Add util to retrieve YAMA ptrace scope
Summary: YAMA is one Linux Security Module, installed and enabled by default on Ubuntu (and maybe other distros), that is particularly relevant for CMA since all it does is restrict ptrace (which is what CMA uses). It can do so at various levels, and the current level can be retrieved from a procfs file. So here I add an utility to do just that.

Reviewed By: beauby

Differential Revision: D25884023

fbshipit-source-id: c21c8e0576c8943051c5b4cee1057836c28d8628