Improve descriptor and viability check for SHM
Summary:
Here I introduce a viability check for SHM, and improve its domain descriptor.

I believe that often those two values are quite related. (Maybe not so much here, but you'll see that in CMA they have a lot in common). Thus the code looks more readable to me if they are calculated together, in the same function. (This is also less error-prone). However, due to the fact that the descriptor must be passed by the context's constructor to the base class's constructor means we cannot call this function in the context's constructor. Instead we'll need a factory method.

This is somewhat annoying but, first of all, it doesn't change the public user interface and, second, it kinda pushes us into a direction we had started to consider anyways, namely the idea of not making a transport's context class public. This is because such a class will never (*) have any custom methods, since it will be immediately down-cast to a generic context pointer. Hence to avoid having to define all these identical context classes we could just have each transport define a static factory function that returns an already-downcast pointer.

Moreover, this new factory method I'm introducing here made me realize that perhaps we might want to change how we handle viability: right now it's an attribute of the transport/channel that is read by the core context, but the user never really becomes aware of it. What if, once we had that factory function I just mentioned, we simply returned a nullptr whenever a viability check failed? This would avoid the attribute and would allow the user to detect and potentially deal with non-viability.

Reviewed By: beauby

Differential Revision: D25884028

fbshipit-source-id: 8998dfce2b07573844c6bb7655bb8c029f93c0a2