Add util to retrieve active Linux Security Modules
Summary: CMA needs ptrace permissions, which involve Linux Security Modules (LSMs), thus in order to determine in advance whether CMA will work we should in principle see what LSMs are active (and then determine whether any of these LSMs would block our call). This may however be a bit overkill, since we won't be able to support all LSMs, and in fact in this first version we'll support only two. However, since it's not too hard to get the full list, I think it's a nice thing to do, as it allows us to then log this information which will make it easy to retrieve from users when we're debugging an issue about this.

Reviewed By: beauby

Differential Revision: D25884024

fbshipit-source-id: af385fd4dd9df905fc509b4bb5ae2e1dec37d3ab