Add util to extract ID for set of permitted Linux capabilities
Summary:
Another LSM that affects CMA is the capability (a.k.a., commoncap) LSM, which is builtin to every kernel and always enabled. In order to check whether it would allow a ptrace between two processes we need to compare their capabilities, thus here I add a utility to query those.

Capabilities are a mess. The "supposed" way to access them "easily" is through libcap, which however in our case adds a bunch of issues: an extra dependency (with thus extra linking issues) and, more importantly, an extra source of mismatch between libcap version and kernel version that could give us bad results (more on this in a sec). There is of course also a real syscall, which is what libcap uses internally, which is rather cumbersome to use. However, that seems to be our best choice, so that's what I am going for. That syscall seems to be properly wrapped by libc, but I couldn't find any header in which it's declared. Hence I ended up doing the weird thing of forward-declaring it (as extern) in our own code. All works fine, and this is also the way libcap does it, so maybe it's the "supposed" way?

One more comment about capabilities: I have the impression that all the APIs around them are designed to get (and set) one _specific_ capability, that is "hardcoded" into an application (e.g., it allows applications to query whether they have CAP_SYS_PTRACE), probably based on the assumption that an application doesn't care about querying capabilities that it doesn't know exist. However, that's precisely what we want to do! We want all capabilities, and we don't care what any of them means. This is why a libcap version mismatch is problematic, as it artificially restricts the queryable capabilities to a subset, and this is also why the syscall itself isn't a perfect solution: once there will be more than 64 capabilities (which won't be the case for many years, given the current pace), all the excess capabilities won't be returned to us, unless we upgrade our code. This is probably by design, and I don't see any way around it.

Reviewed By: beauby

Differential Revision: D25884025

fbshipit-source-id: aec223a0dc4c03bb85d8b5d467fef30d97e89adc