Add viability check to cuda_xth.
Universal addressing is required for cross-device `cudaMemcpyAsync()`,
so this check ensures all visible devices have UVA enabled. This is
the case for all recent devices (i.e. compute capability >=2.0) on
64bit processes.
Should we want to support 32bit processes or very old
hardware (pre 2011 from what I gathered), we could add a fallback to
`cudaMemcpyPeerAsync()` which supports passing the devices as parameters.