Add tests for CMA autodetection (#271)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/271

The goal of these tests is to verify that the viability and the domain descriptor of the CMA channel are able to capture whether running the process_vm_readv syscall would succeed. (False negatives are tolerable, false positives are baaad!). Hence the tests do exactly that: through the use of two "probe" processes, we first ask them to each make a process_vm_readv syscall against the other and report the result, and then also report the viability and the domain descriptor of CMA. The goal is to then launch these processes in different circumstances and validate their reports.

The processes use a UNIX domain socket for setup, mainly because that's the only tool I found that was able to translate PIDs across namespaces (i.e., each probe needs the other one's PID but in its own PID namespace).

Knowing that Docker is a commonly used tool where CMA often fails, I added a Bash script that sets up various "common" Docker setups and launches the probes in them.

Reviewed By: beauby

Differential Revision: D25947731

fbshipit-source-id: d0271b90945ff9e2a0ff7b5020b641bb0de36781