Fix deadlock on destruction of non-viable IBV
Summary:
When the IBV reactor detected, in its constructor, that it wasn't viable it would return immediately and not even start its thread. This left the event loop superclass in an odd state, where it though the thread was serving deferred functions, whereas it wasn't. Hence those deferred functions never got called, and closing+joining deadlocked.

This fix patches the immediate problem, adding the appropriate checks that allow the thread to never be started. They work, but the long-term fix could be another one: instead of supporting two states inside the reactor (viable vs non-viable), we should just avoid creating the reactor if it's not viable. This is a tendency we've started elsewhere, in which we have a factory method determine viability, and which could thus just return a nullptr in case of problems.

Reviewed By: beauby

Differential Revision: D26072723

fbshipit-source-id: 6938a2f34039be83d4fdc302b924eeaf75131cf2