Ensure we call cuInit as soon as we load libcuda
Summary: I was playing with libcuda through Python with ctypes and got some "not initialized" errors. I found out one is supposed to call cuInit before using libcuda, and indeed that fixed my errors. I suspect for the most part we don't need it in practice because the CUDA runtime does this for us, but there might be chances where we use the driver first (especially now that we're doing more and more stuff upon context initialization) hence let's play it safe and call cuInit ourselves. The doc is scarce on this, but it doesn't seem to be a problem to call it multiple times, nor is there a "deinit" function.

Reviewed By: beauby

Differential Revision: D26046667

fbshipit-source-id: 0b165ac22b1368c0e1a32477a65a543e9e2e7867