Provide helper to load NVML (a.k.a., libnvidia-ml.so)
Summary: We'll need NVML for CUDA IPC viability and domain descriptors. Here I add the scaffolding to load it, and a few basic functions. More functions will be added as needed. The whole thing is basically identical to what we already have for libibverbs and libcuda. One minor catch is that NVML has both a nvmlInit and a nvmlShutdown function, which we must call correctly (not too many times, not too few) as they control a reference count which matters. Hence we add a RAII behavior to NvmlLib for that.

Reviewed By: beauby

Differential Revision: D26046668

fbshipit-source-id: 3dc11455d9d11c4603ff0c447a004e3c7a4cf307