Allow transports and channels to provide rich descriptor comparator
Summary:
For CUDA IPC we've found out we need a very specific domain descriptor comparator (pairwise P2P check for all visible devices of the two endpoints), which AFAICT cannot fit into the simple string comparison framework we had used so far. That's a shame, but it's very easy to generalize that check and provide a "rich" comparison.

The main drawback is that we lose the explicit guarantee of symmetry of the comparison. We could refactor the handshake code to only use a backend if _both_ sides agree it can be used, instead of just asking the receiver as we do now. But this would make the handshake much more complicated (and slower), with no real benefit: for the IPC use case, one endpoint can determine "bidirectional" availability on its own. So instead in the following diff I'll just add a safety check, which will raise an error if we detect that the symmetry of the comparison was violated.

Reviewed By: beauby

Differential Revision: D26045364

fbshipit-source-id: c95d1bae79930534ea6e0acf76c60eafacad6e97