Avoid PrivateIfaces in core by declaring Impls in headers
Summary:
A lot of "goodness" has been added to transports and contexts (like the better lifetime management with enrollment, which killed closing emitters and receivers). It's time to bring this to the core pipes too.

The first step is to align the class structures, by removing private ifaces and instead using just impls, defined in their own headers. This is done in this diff, which is massive, but should introduce no functional changes, it just moves code around.

Reviewed By: beauby

Differential Revision: D26045362

fbshipit-source-id: b9fab793cf8884a7804db82f530bec9a9c3364d7