Fix CircleCI tests (#284)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/284

Our CircleCI machines seem to be configured so that `libnvidia-ml.so.1` can be resolved but `libnvidia-ml.so` can't. I don't know if this is a common setup, but fixing it seems simple enough. If we ever find a situation where the opposite happens we'll need to fix our code to attempt both names.

Reviewed By: beauby

Differential Revision: D26100859

fbshipit-source-id: 889f98848e0e07df18bcbb29af622cac69e4991b