Make Pipe provide multiple connections to channels. (#285)
Summary:
With this change, channel contexts define a `numConnectionsNeeded()` method, which is used by the Pipe to provide the requested number of control connections to the `createChannel()` call, which now takes a vector of connections as argument.
This change is needed in order to introduce a CUDA host memory allocator in CudaBasic to avoid costly `cudaMallocHost()`/`cudaFreeHost()` calls (which can hang under certain circumstances).

The changes in actual channels are all the same: forwarding the `numConnectionsNeeded()` call to the impl. The changes to the tests are all the same: wrapping the (single) connection passed to `createChannel()` into a vector. The main changes are in core/pipe_impl.{h,cc}, core/nop_types.h, channel/context.h, channel/context_impl_boilerplate.h.

Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/285

Reviewed By: lw

Differential Revision: D26127334

Pulled By: beauby

fbshipit-source-id: 6c2113b2b6f33044f4085c5b53b84379642730ac