Don't refer to UV's Loop in the libuv helpers
Summary:
There was an odd (almost-)cyclical reference between the libuv helpers (in uv.h) and the UV Loop. It didn't have to be, and it can be easily cut: the libuv helpers only need the uv_loop_t pointer (once) and a DeferredExecutor (many times). So let's specify those directly.

I am making the inLoop of DeferredExecutor const-qualified, so that the libuv helpers can get a const ref to the DeferredExecutor, which I thought was a nice touch. (And it will help later).

Reviewed By: beauby

Differential Revision: D25061005

fbshipit-source-id: ff42e6a1d5786567b53484e0da58b8352b5984d3