Introduce factory functions for channels
Summary:
We want to remove the context definition that are duplicated for all channels, because they do nothing but forward calls to the impls. However these classes need to exist and be visible to users as that's how users create an instance of a certain channel.

Yet, users don't really care to know what exact class a channel context uses: all they need to know about a channel type is that it implements the abstract interface, and how to get hold of an instance of it. Hence a factory function that would return a pointer that is already downcast to the superclass would do just fine. It would allow us to replace the many context classes with a templated boilerplate.

Such a transition needs to happen in phases, because it changes the public API and thus the migration of PyTorch across both GitHub and the internal repo needs to proceed carefully. This is phase 1, where we introduce the new API while keeping the old. Then we migrate use cases. Then we remove the old API.

Reviewed By: beauby

Differential Revision: D26147947

fbshipit-source-id: 4cf11947484f9be358aa83303aa4645b7407a062