Require each channel to implement isViable
Summary: This basically consists in _not_ providing a default implementation of isViable in the abstract base class, and instead require each context to implement it. This is an intermediate step, with the final goal being that the boilerplate handles this. There isn't a strong reason for that, except uniformity as it allows us to deal both with always-viable and with sometimes-viable channels in the same way.

Reviewed By: beauby

Differential Revision: D26102747

fbshipit-source-id: fb4821f41503e87b7a55658f5164ce8084d9bf42