Use factory method in CUDA GDR
Summary:
Factory methods have proven to be more readable than doing the same in the constructor (e.g., we ended up using them both in CMA and CUDA_IPC), moreover they will make it easier for us to pass the viability bit to the boilerplate's constructor.

CUDA_GDR was the channel with most complicated viability check that wasn't yet using a factory method, so here we go.

Reviewed By: beauby

Differential Revision: D26102746

fbshipit-source-id: fed003c2bbcf9b323bd300cf57f6628dff4510d0