Avoid dangling references in CallbackWrappers
Summary:
The CallbackWrapper classes (both the lazy and the eager ones) captured a smart pointer to the object in the closure of the lambda that they return (respectively a weak_ptr and a shared_ptr, but the weak one is locked back to a shared_ptr when the lambda executes). However, the body of that lambda then in turn deferred a task to the loop but that task only captured a (non-owning) _reference_ to the object. Hence by the time that deferred task actually ran in the loop the original smart pointer was gone and thus nothing was guaranteeing that the object was still alive. This lead to heap-use-after-free errors in ASAN. The solution is easy: we can move the owning shared_ptr into the deferred task, so that the object will be kept alive.

I'm quite surprised by this bug, as this is code that we have had around, unchanged, for quite a while, and never experienced problems with it. This bug got triggered quite a bit when I tested the following diff in this stack. That, in fact, can explain why we never saw it in the "core" objects: until now they each had their own loop, hence deferring a task to the object's loop implicitly ensured the object was alive as long as the loop was alive. In the following diff we merge all the objects' loops into a single one, hence the lifetime of objects could become shorter than the one of the loop we defer too. While this explains why core objects didn't see this bug, I still haven't understood why channels (which are the other big users of CallbackWrappers) didn't see it.

Reviewed By: beauby

Differential Revision: D26174243

fbshipit-source-id: daa8df8659232f8783ad55032ec66acd189546d5