Make pipes and listeners share loop with context (#290)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/290

This change is somewhat controversial. Each pipe is independent from each other (to some degree), and quite lightweight: all it does is bookkeeping, it doesn't own resources. The core context doesn't have its own thread(s), hence we use on-demand loops, and we used to have one per pipe (and per listener) so that they could all operate in parallel (provided there were enough "real" underlying event loop to power them).

However, pipes and listeners and their context do sometime interact, for example when initing and closing. And at those times, their lack of synchronization could cause problems. For example when the context would shutdown it would try to close in turn the listeners and the pipes, but this would just enqueue a close operation in their loops, which could be delayed and in the meantime they could attempt to execute another operation which might fail if it tried to access the context as it is already closed.

I'm not currently aware of any concrete issue of that type in the wild, but we did have similar real issues with transports and contexts when they followed a similar design. We since fixed those, by having their contexts immediately (synchronously) set their objects into error, and other such fixes. In this stack I want to do the same for the core context and pipes, and this diff is a prerequisite in order to be able to do so: it makes sure that _all_ pipes and listeners of a certain context use a _single_ shared on-demand loop.

The key upside of it is a greater ease of synchronization between objects. The downside is _potential_ worse performance as this centralized loop could act as a bottleneck. This is what's controversial about this diff. I'll try to argue for it in a couple of ways. First of all, the operations of pipes and listeners are so lightweight, that I would be surprised if we were bottlenecked by the fact that they cannot run in parallel: I'd expect the fraction of time spent in the pipe to be small compared to the one spent in transports and channels, where the real heavy lifting happens. Second, even if that were the case, here we'd be trading performance for correctness and safety, and that may still be a good exchange to make. Finally, if that really becomes a problem (in actual benchmarks), I think we could come up with ways to get the best of both worlds, although that would require some tricky design and implementation (think of a shared loop which however is not mutually exclusive, like a shared lock: pipes can acquire a read-only lock for tasks that can run in isolation, but a read-write lock for operations that involve the context).

Reviewed By: beauby

Differential Revision: D26172367

fbshipit-source-id: f5cc87fb0f24dd5c110bb6e1684a80dd58f53e61