Update on "Refactor channel client-server tests."
Client/server channel tests now deal only with a pre-established
channel. This allows us to automatically create the required number of
connections per channel.
In order to guarantee the connection order on each side, the client
side send the connection id over the connection so that the server
side can match it properly.

[ghstack-poisoned]