Remove last usages of weak_ptr
Summary:
I think our latest "guidance" is that weak_ptr is considered harmful, due to its unpredictability and it encouraging "careless" practices (i.e., not requiring one to think about object lifetime and cleanup). The two remaining instances (*) of weak_ptrs were found in the two places where we accept and dispatch incoming connections, namely the core listener and the MPT context. There was actually no need for their use, because after the previous diff (D26275200) we now ensure that those pending connections are closed during shutdown, and thus that the callbacks will be fired.

(*) There is one (and only one!) other weak_ptr used in a test. It would be nice to get rid of that too, but it's less important.

Reviewed By: beauby

Differential Revision: D26275197

fbshipit-source-id: 782fa66f7f7d3d644923768066a8f6a110a6393c