init listener impl boilerplate with runInLoop
This fixes a bug in core/listener_impl.cc where:
context->listen(address) does a deferred init of the listener, but listener->addr()) 2 lines down does a runInLoop addr() on the same listener. When this is run from within the loop itself, init will be deferred but addr will be called immediately. addr will then throw an exception since listener has not been initialized.
This fix simply makes init use runInLoop too, which makes it initialize immediately when run from within the loop.
Note that the code usually works since new listeners are usually not created from within the loop, but I don't see why this should not be allowed