Introduce async init for contexts
Summary: The context of the MPT channel is managing some objects itself (namely some listeners), hence it needs to register callbacks on them, and it needs to capture a shared_ptr to itself when it does so. It cannot do it from its constructor, hence it has an `init` method for that. (The ContextImpl has it, and the Context factory calls it). This seems like it could one day become handy for other contexts too, so in order to remove any specific quirks from MPT and to future-proof our boilerplate I'm adding this to the generic helpers. So far it's a no-op everywhere except for MPT.

Reviewed By: beauby

Differential Revision: D26337124

fbshipit-source-id: 6b1cf68ab3f970541533a05fa76ba9e97f3f2299