Put non-viable contexts into error state
Summary: Non-viable contexts don't work and thus shouldn't be used. Until now this wasn't enforced, and I guess that attempting to use them could result in nasty errors (segfaults? deadlocks?). It makes sense to close a context when it's non-viable, so that it's effectively unusable. As closing a context now has become the same as setting it to error, we can even use a dedicated error type for it, for extra clarity.

Reviewed By: beauby

Differential Revision: D26337123

fbshipit-source-id: bad7b74e629f71fc0433d01923327dbfe44a03c3