Introduce error state for contexts
Summary:
For pipes, connections, channel, ... (all "child" objects), closing is nothing more than transitioning to the error state. Contexts however were doing it differently, didn't have an error and were using a boolean flag. This is because typically contexts don't perform actions on their own and thus cannot "fail" and don't need an error state. However, one context did: the MPT channel. That's because that context was also doubling as a listener-multiplexer of sorts, hence was directly issuing accept requests. It thus had its own error state, which was separate from the closing state, but somewhat tied.

Moreover, the way the closing state was handled by all contexts was very similar to how child objects dealt with error states: it could only be modified from within the loop (hence closing was deferred), it wouldn't be overridden if it was already set, upon first occurrence it would trigger the cleanup, ...

I think it makes sense to change the closing mechanism in all contexts to be based on an error state, in order to align it with the child objects and to merge MPT's custom logic into the generic framework.

I'm doing it here for transports, channels and core layer all at once.

Reviewed By: beauby

Differential Revision: D26337122

fbshipit-source-id: 9afe231445a8dba20629c7275a9c50da120459be