Add callback wrapper to contexts
Summary: This is another simple future-proofing. We just added the ability to contexts of performing an async init and of handling an error state. These typically go hand-in-hand with the context launching itself some operations on child objects, for which it would probably need a callback wrapper. So let's put the callback wrapper on the boilerplate (and thus remove it from MPT, which is the only channel that currently needs it).

Reviewed By: beauby

Differential Revision: D26337121

fbshipit-source-id: 3afe66d9f12832e099194269e7d9ec3924ffbf6a