Make CUDA multi-device tests harder (#297)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/297

Originally, some of the device guards I added in an earlier diff weren't in fact necessary, which I found odd. Then I realized the tests would actually leave the current device set when calling TensorPipe. This doesn't have to be the case, so I changed the tests to reset the current device to 0, and that did it: now every CUDA call needed a guard.

Differential Revision: D26449415

fbshipit-source-id: 74760f3c3d5b7a2311b5e4360fc9e5001bf96e96