Log full filename of dlopened shared objects
Summary: We're having a bunch of issues where the shared objects we're dynamically opening at runtime aren't the ones we expect (e.g., loading the CUDA driver stub instead of the real thing). To make these issues easier to debug, I'm adding the option to log the full filename of the shared object we've loaded.

Reviewed By: beauby

Differential Revision: D26400685

fbshipit-source-id: bca7da2f6086294d402e029d485205f1d8830a24