Fix longstanding typo in CMA
Summary: The SystemError class expects to take the name of the syscall that returned the error. However, for the `process_vm_readv` syscall we were passing `cma` for that argument. I think this goes back to when CMA was first introduced, as the working name was `process_vm_readv` and perhaps I have been a bit too zealous in my find-and-replace when I renamed it to `cma`.

Reviewed By: beauby

Differential Revision: D26400684

fbshipit-source-id: fd4f36941344e20b003936d990e0976cad6c37e1