Make CUDA multi-device tests harder (#297)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/297

Originally, some of the device guards I added in an earlier diff weren't in fact necessary, which I found odd. Then I realized the tests would actually leave the current device set when calling TensorPipe. This doesn't have to be the case, so I changed the tests to reset the current device to 0, and that did it: now every CUDA call needed a guard.

Reviewed By: beauby

Differential Revision: D26449415

fbshipit-source-id: a1b7f61b225ab0c1bc1e6033b5a09705f871d6aa