Add extra device guards to avoid initializing extra contexts
Summary: We got reports from PyTorch that TensorPipe was initializing CUDA contexts on device 0 even when used exclusively on other devices. I added a check for that and made our tests "stricter" (this will be found in later diffs) and indeed saw that problem. I added extra device guards and fixed the device-inference function, and that fixed them. Note that now every CUDA call has a device guard, and I think that in order to keep this problem popping up again we should make that a habit. (The tests will also help ensure we don't regress on this).

Reviewed By: beauby

Differential Revision: D26423667

fbshipit-source-id: 23e075d01327c50acc66c6462c93130a95a396bc