Move helpers to common files
Summary:
We started having some string helpers duplicated in a couple of places, so it's time to start having a string formatting header.

Also, I'll soon need to access the UUIDs of the visible GPUs in the tests, hence I'm also moving the helper for that from CUDA IPC to a common file.

Reviewed By: beauby

Differential Revision: D26423668

fbshipit-source-id: 5131088c87912b87244cc42645e6cc3ad08e8085