Ensure CUDA tests don't open contexts on unexpected devices
Summary: Add a helper to our CUDA multi-device tests that uses NVML to obtain the list of "active" processes on a given GPU and check if the caller's PID is in there. Then use that helper at the end of each test, on both endpoints, to ensure only the expected CUDA contexts were initialized.

Reviewed By: beauby

Differential Revision: D26423665

fbshipit-source-id: edec899e054dcdb612eb1758a3d73a81af2b14da