Fix synchronization in CUDA tests. (#302)
Summary:
Pull Request resolved: https://github.com/pytorch/tensorpipe/pull/302

For "generic" channel tests, we wrap an std::vector into a
tensorpipe::CudaBuffer by allocating a CUDA buffer of the same size
and calling cudaMemcpy(). We also create a non-blocking stream for
this buffer's use.
The issue here is that the stream being non-blocking, it can advance
before the copy actually starts, which manifested in the following
diff. Replacing the cudaMemcpy with a cudaMemcpyAsync enqueued on the
stream fixes the issue, and corresponds to the intended use of CUDA
buffers in TensorPipe. I am not certain why it hasn't manifested in
tests for other CUDA channels though.

Test Plan: Imported from OSS

Reviewed By: lw

Differential Revision: D26575582

Pulled By: beauby

fbshipit-source-id: e03f2e5c79818389f43ccf7debfec71b716100e6