Trim filenames in logging
Summary: In our logs we include the filename of where the log originated, but we use the compiler's "version" of the filename, which could be the full explicit path. This typically adds noise, as we don't care where that file is in the filesystem, we just want its "logical" name. In this diff, I add some code to "strip" the leading part of filenames before we log them.

Reviewed By: beauby

Differential Revision: D26604487

fbshipit-source-id: eb56835c947066b146a16040c17e65df174cef20