Deduplicate member field in pipe
Summary: This seems unrelated to the rest of the stack, but in the end it is: in order to track the range of read operations for which we are expecting a read call, we used to store the sequence numbers of the endpoints of that interval. I want to get rid of that, and opt to use iterators rather than sequence numbers. In this diff I get rid of storing one of the endpoints of that interval, which turned out to be useless as we were already keeping track of it for other purposes.

Reviewed By: beauby

Differential Revision: D26223399

fbshipit-source-id: f4d1494e6ea5a1d893d4390ed30ba34121cce148