Remove implicit cross-op monotonicity from state machine
Summary: The state machine, when it was used in the pipe, implicitly ensured that later operations could not "overtake" earlier ones. This was how the pipe wanted it work, but in other cases we might actually want such a behavior, hence here I relax such behavior from the reusable component and instead move it to the pipe's transition table.

Reviewed By: beauby

Differential Revision: D26223397

fbshipit-source-id: 35e43d20f7ff8c8db9623ada4ddc3eb86a29a42c