Use iterators in state machine's interface
Summary:
Having the state machine deal with references or pointers to the operations is limiting, because in order to operate on the container (erase elements, move from one element to the next, ...) one needs iterators, and there is no way to obtain the iterator from a pointer or reference (except, as we did here, for getting the sequence number and then finding it in the container). On the other hand, the reverse (getting the pointer or reference given the iterator) is trivial.

In order to have greater liberty in our implementation (e.g., to use a list instead of a deque, in which case we wouldn't be able to efficiently look up an operation given its sequence number) it makes more sense to use iterators everywhere. In this diff, I change _only_ the API of the state machine (and thus all its callsites) while leaving the internal logic of the state machine entirely unchanged.

Reviewed By: beauby

Differential Revision: D26223398

fbshipit-source-id: bcb527fbfadb3a37d4ff4d0970b4cc24aabe999e