Don't risk accessing op after it's destroyed in CUDA GDR
Summary: This fixes another minor bug in CUDA GDR. We don't wait for the ready-to-receive control message to flush before destroying the recv operation, because there's nothing we need to do once it happens. However, this means that its callback couldn't safely access the operation, as it could have already been destroyed. This was only happening when debug logging was enabled, as we were reading the sequence number. To avoid problems, let's just "unpack" the sequence number when building the lambda, thus avoiding any risk.

Reviewed By: beauby

Differential Revision: D26460334

fbshipit-source-id: 2f247375d9064420c16904846f144c21a9b1002b