Stop pipe querying ops in state machine by seq number
Summary:
The state machine helper exposes a method for accessing an operation given its sequence number. This is only used by the pipe in one occasion, namely to retrieve the (already existing) read operation for which the user is calling the read method (this must happen after we've invoked the user's readDescriptor callback for that operation). Access by sequence number can only be done efficiently if the operations are stored in contiguous memory, like a vector or a deque, and would thus hinder us from using for example a list or other data structures which might be more appropriate once we allow later operations to finish before earlier ones. Thus, in order to have the ability to adapt our implementation, it would be nice if we could "narrow" our API and remove this method.

In this diff I remove that usage of access-by-sequence-number by the pipe. Note that, under "normal" operation, there is at all times either one or zero read operations for which we're expecting a read call from the user. (That's because in order to read the descriptor of the next operation, and thus invoke the readDescriptor callback, we first must read the payloads of the previous operation, and we can only do so once the user gave us some buffers by calling read). In such a case, it's easy to solve the problem: when we invoke the readDescriptor callback we store a pointer to the operation, and then we read that pointer when read is called.

However, in error cases, the pipe will flush all readDescriptor callbacks at once, which will lead to a situation of multiple read calls being expected. The above solution won't work, as we'd have to store a queue of pointers. However, that's wasteful (under normal operations) and would in some sense duplicate the existing queue of operations that is already handled by the state machine. Instead, we can solve this problem by leveraging the state machine by introducing a new state, which identifies the _first_ operation for which we're expecting a read call. Thus, there will always be either one or zero ops in that new state, even in case of error, and thus the conditions of our previous solution still apply and storing a single pointer works.

Reviewed By: beauby

Differential Revision: D26223400

fbshipit-source-id: 56f8f4cb77027a0794b0acc89a05bc6a46e74f2f