Use state machine in CUDA GDR
Summary:
This replaces the callback hell / spaghetti code of CUDA GDR with a sane state machine. In order to keep the diff small, I kept each former method and made it become either an action or a reaction of the state machine. Hence there's no changes in the "proper" logic, and all I'm adding here is some scaffolding around that.

Note that since I am keeping the logic unchanged, I'm not yet fixing the issues that the state machine was supposed to address. I'll do that in the following diffs, which will be tiny (that's the whole point of the state machine!).

Reviewed By: beauby

Differential Revision: D26460330

fbshipit-source-id: b38ccd45ea6cec55d67b198da889a894b453af30