Use separate control connections for replies and acks in CUDA IPC
Summary: CUDA IPC needs to send two independent streams of messages in each direction (replies and acks). It used to do so by multiplexing on a single connection, but it was somewhat awkward, and will make it more complicated to onboard this channel onto a state machine. However, we've recently allowed channels to request more than one control connections, and doing so for IPC would greatly simplify the logic. The only "performance" hit is that we open a few more file descriptors, but this has not proven to be a concern yet.

Differential Revision: D26483685

fbshipit-source-id: 079e70622c9ee93edbbb9b3ce4760ab6e0fd0ec6